+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Elgg Cropper - CHANGE LOG
 ### 2.1.1 (2016-11-15)
 * Plenty of Cropper tool options are now available on plugin settings
 * Added option to select size of preview box in settings
+* Aspectratio cropper option can be passed as param on input view, overriding the plugin settings
 
 ### 2.1.0 (2016-11-07)
 * **releases:** Initial release